requirements.txt
Django==2.2.16
pytz==2018.5
psycopg2-binary
python-dotenv
djangorestframework
PyYaml
